requirements.txt
# Book Factory - Python Requirements
# Install with: pip install -r requirements.txt

# ============================================
# Core Dependencies
# ============================================

# API Clients
openai>=1.0.0
anthropic>=0.18.0
groq>=0.4.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Data Processing
pyyaml>=6.0.1
python-dotenv>=1.0.0
jsonschema>=4.20.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# ============================================
# Quality & Analysis
# ============================================

# Text Analysis
nltk>=3.8.1
textstat>=0.7.3
spacy>=3.7.0
textblob>=0.17.1

# Data Science
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0

# ============================================
# Cloud & Infrastructure
# ============================================

# Oracle Cloud
oci>=2.118.0

# AWS (future)
# boto3>=1.34.0

# Monitoring
sentry-sdk>=1.39.0
prometheus-client>=0.19.0

# ============================================
# Publishing & Generation
# ============================================

# Image Generation
pillow>=10.1.0
pillow-heif>=0.14.0
cairosvg>=2.7.1

# PDF Generation
reportlab>=4.0.0
pypdf>=3.17.0

# Markdown Processing
markdown>=3.5.0
markdown2>=2.4.12
mistune>=3.0.0

# ============================================
# Web Framework (Future Dashboard)
# ============================================

# FastAPI
fastapi>=0.108.0
uvicorn>=0.25.0
jinja2>=3.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ============================================
# Development & Testing
# ============================================

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
responses>=0.24.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
pylint>=3.0.0

# Development Tools
ipython>=8.18.0
ipdb>=0.13.13
rich>=13.7.0  # Better terminal output
click>=8.1.0  # CLI framework
tqdm>=4.66.0  # Progress bars

# ============================================
# Utilities
# ============================================

# Async
asyncio>=3.4.3
aiofiles>=23.2.1

# Scheduling
schedule>=1.2.0
apscheduler>=3.10.4

# Caching
cachetools>=5.3.2
diskcache>=5.6.3

# Configuration
configparser>=6.0.0
environs>=11.0.0

# Logging
loguru>=0.7.2
colorlog>=6.8.0

# Rate Limiting
ratelimit>=2.2.1
slowapi>=0.1.9

# Retry Logic
tenacity>=8.2.3
backoff>=2.2.1

# ============================================
# Optional: Advanced Features
# ============================================

# Machine Learning
# transformers>=4.36.0  # Hugging Face
# torch>=2.1.0  # PyTorch
# tensorflow>=2.15.0  # TensorFlow

# Graph Processing (for series coherence)
# networkx>=3.2.0

# Advanced NLP
# sentence-transformers>=2.2.2
# langchain>=0.1.0

# ============================================
# Platform Specific
# ============================================

# Windows specific
# pywin32>=306 ; sys_platform == 'win32'

# Unix specific
# python-daemon>=3.0.1 ; sys_platform != 'win32'